remove impersonation
fixes  #1642
