+++ src/NServiceBus.Core/Impersonation/ExtractIncomingPrincipal.cs
@@ -5,6 +5,12 @@
     /// <summary>
     /// Allows different authentication techniques to be plugged in.
     /// </summary>
+    [ObsoleteEx(
+        Message = "The impersonation feature has been removed due to confusion of it being a security feature." +
+                  "Once you stop using this feature the Thread.CurrentPrincipal will no longer be set to a fake principal containing the username. However you can still get access to that information using the message headers.",
+        Replacement = "message.GetHeader(Headers.WindowsIdentityName)",
+        RemoveInVersion = "5.0",
+        TreatAsErrorFromVersion = "4.3")]
     public interface ExtractIncomingPrincipal
     {
         /// <summary>
