+++ src/NServiceBus.Core/Impersonation/Windows/ConfigureWindowsIdentityEnricher.cs
@@ -1,0 +1,11 @@
+namespace NServiceBus.Impersonation.Windows
+{
+    internal class ConfigureWindowsIdentityEnricher : IWantToRunBeforeConfigurationIsFinalized
+    {
+        public void Run()
+        {
+            Configure.Instance.Configurer.ConfigureComponent<WindowsIdentityEnricher>(DependencyLifecycle.SingleInstance);
+        }
+
+    }
+}
